requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
google-generativeai
elevenlabs
moviepy
pydantic
scipy
asyncpg
sqlalchemy[asyncio]
pyjwt
bcrypt
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2